+++ src/components/checkbox/checkbox.test.ts
@@ -171,17 +171,17 @@ describe('<wa-checkbox>', () => {
 
       expect(checkbox.checkValidity()).to.be.false;
       expect(checkbox.checkValidity()).to.be.false;
-      expect(checkbox.hasAttribute('data-invalid')).to.be.true;
-      expect(checkbox.hasAttribute('data-valid')).to.be.false;
-      expect(checkbox.hasAttribute('data-user-invalid')).to.be.true;
-      expect(checkbox.hasAttribute('data-user-valid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-invalid')).to.be.true;
+      expect(checkbox.hasAttribute('data-wa-valid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-user-invalid')).to.be.true;
+      expect(checkbox.hasAttribute('data-wa-user-valid')).to.be.false;
 
       await clickOnElement(checkbox);
       await checkbox.updateComplete;
       await aTimeout(0);
 
-      expect(checkbox.hasAttribute('data-user-invalid')).to.be.true;
-      expect(checkbox.hasAttribute('data-user-valid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-user-invalid')).to.be.true;
+      expect(checkbox.hasAttribute('data-wa-user-valid')).to.be.false;
     });
 
     it('should be invalid when required and unchecked', async () => {
@@ -214,12 +214,12 @@ describe('<wa-checkbox>', () => {
       const el = await fixture<HTMLFormElement>(html` <form novalidate><wa-checkbox required></wa-checkbox></form> `);
       const checkbox = el.querySelector<WaCheckbox>('wa-checkbox')!;
 
-      expect(checkbox.hasAttribute('data-required')).to.be.true;
-      expect(checkbox.hasAttribute('data-optional')).to.be.false;
-      expect(checkbox.hasAttribute('data-invalid')).to.be.true;
-      expect(checkbox.hasAttribute('data-valid')).to.be.false;
-      expect(checkbox.hasAttribute('data-user-invalid')).to.be.false;
-      expect(checkbox.hasAttribute('data-user-valid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-required')).to.be.true;
+      expect(checkbox.hasAttribute('data-wa-optional')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-invalid')).to.be.true;
+      expect(checkbox.hasAttribute('data-wa-valid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-user-invalid')).to.be.false;
+      expect(checkbox.hasAttribute('data-wa-user-valid')).to.be.false;
     });
   });