+++ src/components/select/select.test.ts
@@ -302,12 +302,12 @@ describe('<wa-select>', async () => {
       const secondOption = el.querySelectorAll('wa-option')[1];
 
       expect(el.checkValidity()).to.be.true;
-      expect(el.hasAttribute('data-required')).to.be.true;
-      expect(el.hasAttribute('data-optional')).to.be.false;
-      expect(el.hasAttribute('data-invalid')).to.be.false;
-      expect(el.hasAttribute('data-valid')).to.be.true;
-      expect(el.hasAttribute('data-user-invalid')).to.be.false;
-      expect(el.hasAttribute('data-user-valid')).to.be.false;
+      expect(el.hasAttribute('data-wa-required')).to.be.true;
+      expect(el.hasAttribute('data-wa-optional')).to.be.false;
+      expect(el.hasAttribute('data-wa-invalid')).to.be.false;
+      expect(el.hasAttribute('data-wa-valid')).to.be.true;
+      expect(el.hasAttribute('data-wa-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-wa-user-valid')).to.be.false;
 
       await el.show();
       await clickOnElement(secondOption);
@@ -316,8 +316,8 @@ describe('<wa-select>', async () => {
       await el.updateComplete;
 
       expect(el.checkValidity()).to.be.true;
-      expect(el.hasAttribute('data-user-invalid')).to.be.false;
-      expect(el.hasAttribute('data-user-valid')).to.be.true;
+      expect(el.hasAttribute('data-wa-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-wa-user-valid')).to.be.true;
     });
 
     it('should receive the correct validation attributes ("states") when invalid', async () => {
@@ -330,12 +330,12 @@ describe('<wa-select>', async () => {
       `);
       const secondOption = el.querySelectorAll('wa-option')[1];
 
-      expect(el.hasAttribute('data-required')).to.be.true;
-      expect(el.hasAttribute('data-optional')).to.be.false;
-      expect(el.hasAttribute('data-invalid')).to.be.true;
-      expect(el.hasAttribute('data-valid')).to.be.false;
-      expect(el.hasAttribute('data-user-invalid')).to.be.false;
-      expect(el.hasAttribute('data-user-valid')).to.be.false;
+      expect(el.hasAttribute('data-wa-required')).to.be.true;
+      expect(el.hasAttribute('data-wa-optional')).to.be.false;
+      expect(el.hasAttribute('data-wa-invalid')).to.be.true;
+      expect(el.hasAttribute('data-wa-valid')).to.be.false;
+      expect(el.hasAttribute('data-wa-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-wa-user-valid')).to.be.false;
 
       await el.show();
       await clickOnElement(secondOption);
@@ -344,8 +344,8 @@ describe('<wa-select>', async () => {
       el.blur();
       await el.updateComplete;
 
-      expect(el.hasAttribute('data-user-invalid')).to.be.true;
-      expect(el.hasAttribute('data-user-valid')).to.be.false;
+      expect(el.hasAttribute('data-wa-user-invalid')).to.be.true;
+      expect(el.hasAttribute('data-wa-user-valid')).to.be.false;
     });
 
     it('should receive validation attributes ("states") even when novalidate is used on the parent form', async () => {
@@ -360,12 +360,12 @@ describe('<wa-select>', async () => {
       `);
       const select = el.querySelector<WaSelect>('wa-select')!;
 
-      expect(select.hasAttribute('data-required')).to.be.true;
-      expect(select.hasAttribute('data-optional')).to.be.false;
-      expect(select.hasAttribute('data-invalid')).to.be.true;
-      expect(select.hasAttribute('data-valid')).to.be.false;
-      expect(select.hasAttribute('data-user-invalid')).to.be.false;
-      expect(select.hasAttribute('data-user-valid')).to.be.false;
+      expect(select.hasAttribute('data-wa-required')).to.be.true;
+      expect(select.hasAttribute('data-wa-optional')).to.be.false;
+      expect(select.hasAttribute('data-wa-invalid')).to.be.true;
+      expect(select.hasAttribute('data-wa-valid')).to.be.false;
+      expect(select.hasAttribute('data-wa-user-invalid')).to.be.false;
+      expect(select.hasAttribute('data-wa-user-valid')).to.be.false;
     });
   });